% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_distribution.R
\name{distribution_cdf}
\alias{distribution_cdf}
\title{Evaluate the empirical predictive cumulative distribution function}
\usage{
distribution_cdf(object, q)
}
\arguments{
\item{object}{A \code{page_predictive_distribution}.}

\item{q}{Numeric thresholds on the distribution's outcome scale.}
}
\value{
Numeric probabilities \eqn{P(Y \le q)}{P(Y <= q)}.
}
\description{
Evaluate the empirical predictive cumulative distribution function
}
