% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_distribution.R
\name{distribution_draws}
\alias{distribution_draws}
\title{Return predictive draws}
\usage{
distribution_draws(object)
}
\arguments{
\item{object}{A \code{page_predictive_distribution}.}
}
\value{
Numeric vector of predictive draws.
}
\description{
Return predictive draws
}
\seealso{
\code{\link[=distribution_weights]{distribution_weights()}} for
  weighted atoms.
}
