% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retired.R
\name{fit_reference_gam}
\alias{fit_reference_gam}
\title{Fit reference GAM and set closures}
\usage{
fit_reference_gam(df, k = 12)
}
\arguments{
\item{df}{data.frame(season, newWeek, y, neg)}

\item{k}{basis dimension for s(newWeek)}
}
\description{
Fit reference GAM and set closures
}
