% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_runtime_helpers.R
\name{get_gam_cls}
\alias{get_gam_cls}
\title{Extract a classifier GAM from various container objects}
\usage{
get_gam_cls(ign_fit_or_gam)
}
\arguments{
\item{ign_fit_or_gam}{A trained classifier model or a container holding one.}
}
\value{
An \pkg{mgcv} \code{gam} or \code{bam} object.
}
\description{
Convenience helper that accepts either:
\itemize{
\item an \pkg{mgcv} \code{gam}/\code{bam} object;
\item a \code{gamm4} fit list with component \code{$gam};
\item a list returned by your \code{fitIgnition()} that contains
  \code{$fits$p_only_week_p$gam}.
}
}
\examples{
\dontrun{
gam_cls <- get_gam_cls(ign_fit)                        # fitIgnition() output
gam_cls <- get_gam_cls(ign_fit$fits$p_only_week_p$gam) # direct
}
}
