% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_g_ref_fun}
\alias{make_g_ref_fun}
\title{Build reference link-scale function from a fitted GAM}
\usage{
make_g_ref_fun(gam_obj, week_grid = 1:52)
}
\arguments{
\item{gam_obj}{a fitted mgcv::gam (or gamm4::$gam) with predictor `newWeek`}

\item{week_grid}{numeric vector of weeks to interpolate over (default 1:52)}
}
\value{
A function that returns logit probability for arbitrary values of u.
}
\description{
Build reference link-scale function from a fitted GAM
}
