% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expert_timing_annotations.R
\name{new_expert_ignition_annotation}
\alias{new_expert_ignition_annotation}
\title{Create one expert decimal ignition annotation}
\usage{
new_expert_ignition_annotation(
  season,
  ignition_week_decimal,
  n_weeks = 52L,
  ignition_interval = NULL,
  annotator,
  annotation_version,
  annotated_at,
  data_snapshot_id,
  positivity_version,
  comment = NULL
)
}
\description{
Create one expert decimal ignition annotation
}
