% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{PAGe-package}
\alias{PAGe}
\alias{PAGe-package}
\title{PAGe: Seasonal Flu Curve Alignment and Forecasting}
\description{
Align partially observed seasonal positivity curves to a learned reference shape via profile likelihood (shift tau and optional dilation delta), then forecast the remainder of the season with sensible prediction intervals and peak timing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lennon-li/PAGe}
  \item Report bugs at \url{https://github.com/lennon-li/PAGe/issues}
}

}
\author{
\strong{Maintainer}: Lennon Li \email{lennon.li@oahpp.ca}

}
\keyword{internal}
