% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page1}
\alias{page1}
\alias{page2}
\title{Gene Expression Signature Similarity}
\usage{
page1(datM, geneV, isRank = TRUE, adjust = "fdr")

page2(datM, gene_up, gene_dn, isRank = TRUE, adjust = "fdr")
}
\arguments{
\item{datM}{A Gene Expression Signature matrix including logFC or rank}

\item{geneV}{a gene symbol name vector}

\item{isRank}{datM is rank or not}

\item{adjust}{p.adjust.methods. see \code{\link[stats]{p.adjust}}}

\item{gene_up}{a up-regulated gene symbol name vector}

\item{gene_dn}{a down-regulated gene symbol name vector}
}
\description{
Gene Expression Signature Similarity based on Parametric Analysis of Gene Set Enrichment.
}
\examples{
data(Psoriasis_Etanercept_LogFC)
datM <- apply(Psoriasis_Etanercept_LogFC, 2, rank)
gene_up <- names(head(sort(Psoriasis_Etanercept_LogFC[,1], decreasing=TRUE), 100) )
gene_dn <- names(head(sort(Psoriasis_Etanercept_LogFC[,1], decreasing=FALSE), 100) )
page1(datM, gene_up)
page1(datM, gene_dn)
page2(datM, gene_up, gene_dn)

}
\references{
1. Kim, Seon-Young, and David J. Volsky. "PAGE: parametric analysis of gene set 
enrichment." BMC bioinformatics 6.1 (2005): 1.

2. Iorio, Francesco, et al. "Discovery of drug mode of action and drug 
repositioning from transcriptional responses." Proceedings of the National 
Academy of Sciences 107.33 (2010): 14621-14626.
}

