% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGE_package.R
\docType{package}
\name{PAGE_package}
\alias{PAGE_package}
\alias{PAGE_package-package}
\title{PAGE}
\description{
Gene Expression Signature Similarity based on Parametric Analysis of Gene Set Enrichment
}
\keyword{package}
