% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{page_validate_kit}
\alias{page_validate_kit}
\title{Validate a PAGe kit}
\usage{
page_validate_kit(...)
}
\arguments{
\item{...}{Arguments passed to \code{validate_page_kit()}.}
}
\value{
Invisibly, the validated kit.
}
\description{
Validate a PAGe kit
}
