% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeasonCurves.R
\name{plotSeasonCurves}
\alias{plotSeasonCurves}
\title{Plot observed vs fitted positivity by season, with ignition week in title}
\usage{
plotSeasonCurves(df, x = "weekF")
}
\arguments{
\item{df}{Data frame containing at least: season, weekF, y, N, fit.
Also needs either:
- iWeek (ignition weekF repeated within season), OR
- ignition (logical TRUE at ignition row)}

\item{x}{Character. X-axis column name (default "weekF").}
}
\value{
A ggplot object.
}
\description{
Plot observed vs fitted positivity by season, with ignition week in title
}
