% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGE_package.R
\docType{data}
\name{Psoriasis_Etanercept_LogFC}
\alias{Psoriasis_Etanercept_LogFC}
\title{Psoriasis_Etanercept dataset.}
\format{\describe{
\item{Psoriasis_Etanercept_LogFC_mat}{A LogFC matrix related with Psoriasis and Etanercept.}
}}
\description{
An example dataset of Psoriasis and Etanercept.
}
\keyword{datasets}
