% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGE.R
\docType{data}
\name{Psoriasis_Etanercept_LogFC_mat}
\alias{Psoriasis_Etanercept_LogFC_mat}
\title{Psoriasis_Etanercept_LogFC_mat}
\description{
A LogFC matrix related with Psoriasis and Etanercept
}
\keyword{datasets}

