% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGE.R
\docType{data}
\name{Psoriasis_Etanercept_LogFC_rank}
\alias{Psoriasis_Etanercept_LogFC_rank}
\title{Psoriasis_Etanercept_LogFC_rank}
\description{
A LogFC rank matrix related with Psoriasis and Etanercept
}
\keyword{datasets}

