% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickenrichmentscore.R
\name{quickenrichmentscore}
\alias{quickenrichmentscore}
\title{quickenrichmentscore}
\source{
http://bioconductor.org/packages/GeneExpressionSignature/
}
\usage{
quickenrichmentscore(S, S1, List)
}
\arguments{
\item{S}{rank of up-regualted genes}

\item{S1}{rank of down-regulated genes}

\item{List}{a rank vector}
}
\value{
similarity based on averaged GSEA Enrichment score
}
\description{
quickenrichmentscore
}
\examples{
data(Psoriasis_Etanercept_LogFC)
dat <- apply(-Psoriasis_Etanercept_LogFC, 2, rank)
SignatureLength <- 250

# caculate similarity only for some columns
sim_mat <- matrix(-2, ncol=ncol(dat), nrow=ncol(dat))
for (i in 11:15) {
    for (j in 1:10) {
        sim_mat[j,i] <- (quickenrichmentscore(which(dat[,j]<=SignatureLength), 
        which(dat[,j]>=nrow(dat)-SignatureLength+1), dat[,i]) + 
        quickenrichmentscore(which(dat[,i]<=SignatureLength), 
        which(dat[,i]>=nrow(dat)-SignatureLength+1), dat[,j]) )/2
    }

}

}
\author{
Yang Cao
}
