% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expert_timing_review_set.R
\name{review_expert_ignition_set}
\alias{review_expert_ignition_set}
\title{Build expert ignition review objects for all seasons}
\usage{
review_expert_ignition_set(
  data,
  seasons = NULL,
  existing_annotations = NULL,
  ...
)
}
\description{
Build expert ignition review objects for all seasons
}
