% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreGsea.R
\name{scoreGsea}
\alias{scoreGsea}
\alias{scoreGseam}
\title{Gene Expression Signature Similarity based on GSEA}
\usage{
scoreGsea(dat, top = 250, type = c("avg", "max"))

scoreGseam(dat, top = 250, type = c("avg", "max"), ncore = 2,
  verbose = FALSE)
}
\arguments{
\item{dat}{gene rank matrix or data.frame}

\item{top}{top genes selected}

\item{type}{avg or max}

\item{ncore}{the number of cores used}

\item{verbose}{verbose}
}
\description{
Gene Expression Signature Similarity based on  Gene Set Enrichment Analysis
}
\examples{
data(Psoriasis_Etanercept_LogFC)
Psoriasis_Etanercept_rank <- apply(-Psoriasis_Etanercept_LogFC, 2, rank)

sim <- scoreGsea(Psoriasis_Etanercept_rank, type="avg")
sim <- scoreGseam(Psoriasis_Etanercept_rank, type="avg", ncore=2)

}
