% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{shared_denominator_correlation}
\alias{shared_denominator_correlation}
\title{Shared-denominator multinomial correlation}
\usage{
shared_denominator_correlation(...)
}
\arguments{
\item{...}{Arguments passed to \code{page_shared_denominator_correlation()}.}
}
\value{
A stratum correlation matrix.
}
\description{
Shared-denominator multinomial correlation
}
