% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expert_timing_annotations.R
\name{validate_expert_ignition_annotation}
\alias{validate_expert_ignition_annotation}
\title{Validate an expert ignition annotation}
\usage{
validate_expert_ignition_annotation(x)
}
\arguments{
\item{x}{A \code{page_expert_ignition_annotation_v2} object.}
}
\value{
Invisibly, the validated annotation.
}
\description{
Validate an expert ignition annotation
}
