% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorePage.R
\name{zscore_sim}
\alias{zscore_sim}
\title{zscore_sim
convert zscore to similarity}
\usage{
zscore_sim(Zscore, type = "avg", similarity = TRUE, adjust = "fdr")
}
\description{
zscore_sim
convert zscore to similarity
}
\keyword{internal}
