% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expert_timing_review_set.R
\name{expert_ignition_annotation_sheet}
\alias{expert_ignition_annotation_sheet}
\title{Create a blank ignition annotation sheet}
\usage{
expert_ignition_annotation_sheet(
  data,
  annotator,
  annotation_version,
  positivity_version
)
}
\description{
Create a blank ignition annotation sheet
}
