% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expert_timing_workflow.R
\name{finalize_expert_ignition_review}
\alias{finalize_expert_ignition_review}
\title{Finalize numeric expert ignition after review}
\usage{
finalize_expert_ignition_review(
  review,
  ignition_week_decimal,
  annotator,
  annotation_version,
  positivity_version,
  ignition_interval = NULL,
  comment = NULL,
  annotated_at = Sys.time()
)
}
\description{
Finalize numeric expert ignition after review
}
