% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorePage.R
\name{scorePage}
\alias{scorePage}
\title{scorePage: Gene Expression Signature Similarity based on Parametric Analysis of Gene Set Enrichment}
\usage{
scorePage(logFCmat, top = 250, center = TRUE, type = c("avg", "max"),
  similarity = TRUE)
}
\arguments{
\item{logFCmat}{logFC matrix.}

\item{top}{top genes selected.}

\item{center}{boolean - median center gene expression matrix columns prior to analysis.}

\item{type}{avg or max}

\item{similarity}{boolean}
}
\value{
a list
}
\description{
Gene Expression Signature Similarity based on Parametric Analysis of Gene Set Enrichment
}
\examples{

data(Psoriasis_Etanercept_LogFC_mat)
sim <- scorePage(Psoriasis_Etanercept_LogFC_mat, type="avg")


}

