% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorePage.R
\docType{methods}
\name{scorePage}
\alias{scorePage}
\alias{scorePage,matrix-method}
\alias{scorePage}
\alias{scorePage,list-method}
\alias{scorePage}
\title{scorePage}
\usage{
scorePage(dat, top = 250, ncore = 2, type = c("avg", "max"),
  similarity = TRUE, adjust = "fdr", center = TRUE, verbose = FALSE)

\S4method{scorePage}{matrix}(dat, top = 250, ncore = 2, type = c("avg",
  "max"), similarity = TRUE, adjust = "fdr", center = TRUE,
  verbose = FALSE)

\S4method{scorePage}{list}(dat, top = 250, ncore = 2, type = c("avg",
  "max"), similarity = TRUE, adjust = "fdr", center = TRUE,
  verbose = FALSE)
}
\arguments{
\item{dat}{logFC matrix or list}

\item{top}{top genes selected}

\item{ncore}{the number of cores used}

\item{type}{avg or max}

\item{similarity}{boolean}

\item{adjust}{p.adjust.methods. see \code{\link[stats]{p.adjust}}}

\item{center}{boolean}

\item{verbose}{verbose}
}
\value{
a list with score, pval, fdr
}
\description{
Gene Expression Signature Similarity based on Parametric Analysis of Gene Set Enrichment
}
\examples{

data(Psoriasis_Etanercept_LogFC)


sim <- scorePage(Psoriasis_Etanercept_LogFC, type="avg")$score


}
\references{
1. Kim, Seon-Young, and David J. Volsky. "PAGE: parametric analysis of gene set 
enrichment." BMC bioinformatics 6.1 (2005): 1.

2. Iorio, Francesco, et al. "Discovery of drug mode of action and drug 
repositioning from transcriptional responses." Proceedings of the National 
Academy of Sciences 107.33 (2010): 14621-14626.
}
