% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expert_timing_workflow.R
\name{write_expert_ignition_annotations}
\alias{write_expert_ignition_annotations}
\title{Save versioned expert ignition annotations}
\usage{
write_expert_ignition_annotations(annotations, path, overwrite = FALSE)
}
\description{
Save versioned expert ignition annotations
}
